@@ -1,127 +1,104 @@
 <?php
+
 declare(strict_types=1);
+
 namespace YoastSeoForTypo3\YoastSeo\Form\Element;
 
 use TYPO3\CMS\Backend\Form\AbstractNode;
-use TYPO3\CMS\Backend\Form\NodeFactory;
 use TYPO3\CMS\Backend\Utility\BackendUtility;
-use TYPO3\CMS\Core\Authentication\BackendUserAuthentication;
 use TYPO3\CMS\Core\Domain\Repository\PageRepository;
-use TYPO3\CMS\Core\Page\PageRenderer;
-use TYPO3\CMS\Core\Utility\ArrayUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Fluid\View\StandaloneView;
 use TYPO3\CMS\Frontend\Page\CacheHashCalculator;
-use YoastSeoForTypo3\YoastSeo\Service\LocaleService;
+use YoastSeoForTypo3\YoastSeo\Service\SnippetPreviewService;
 use YoastSeoForTypo3\YoastSeo\Service\UrlService;
-use YoastSeoForTypo3\YoastSeo\Utility\JsonConfigUtility;
-use YoastSeoForTypo3\YoastSeo\Utility\PathUtility;
 use YoastSeoForTypo3\YoastSeo\Utility\YoastUtility;
 
 class SnippetPreview extends AbstractNode
 {
-    /**
-     * @var StandaloneView
-     */
-    protected $templateView;
+    protected string $titleField = 'title';
+    protected string $pageTitleField = 'title';
+    protected string $descriptionField = 'description';
+    protected string $focusKeywordField = 'tx_yoastseo_focuskeyword';
+    protected string $focusKeywordSynonymsField = 'tx_yoastseo_focuskeyword_synonyms';
+    protected string $cornerstoneField = 'tx_yoastseo_cornerstone';
+    protected string $relatedKeyphrases = 'tx_yoastseo_focuskeyword_related';
+    protected string $table = 'pages';
+    protected string $previewUrl = '';
+    protected int $languageId = 0;
 
-    /**
-     * @var string
-     */
-    protected $titleField = 'title';
+    protected UrlService $urlService;
 
-    /**
-     * @var string
-     */
-    protected $pageTitleField = 'title';
-
-    /**
-     * @var string
-     */
-    protected $descriptionField = 'description';
-
-    /**
-     * @var string
-     */
-    protected $focusKeywordField = 'tx_yoastseo_focuskeyword';
-
-    /**
-     * @var string
-     */
-    protected $focusKeywordSynonymsField = 'tx_yoastseo_focuskeyword_synonyms';
-
-    /**
-     * @var string
-     */
-    protected $cornerstoneField = 'tx_yoastseo_cornerstone';
-
-    /**
-     * @var string
-     */
-    protected $relatedKeyphrases = 'tx_yoastseo_focuskeyword_premium';
-
-    /**
-     * @var string
-     */
-    protected $table = 'pages';
-
-    /**
-     * @var string
-     */
-    protected $previewUrl = '';
-
-    /**
-     * @var int
-     */
-    protected $languageId = 0;
-
-    /**
-     * @var \YoastSeoForTypo3\YoastSeo\Service\LocaleService
-     */
-    protected $localeService;
-
-    /**
-     * @var \YoastSeoForTypo3\YoastSeo\Service\UrlService
-     */
-    protected $urlService;
-
-    /**
-     * @var array
-     */
-    protected $configuration = [
-        'translations' => [
-            'availableLocales' => [],
-            'languageKeyToLocaleMapping' => []
-        ],
-    ];
-
-    /**
-     * @param NodeFactory $nodeFactory
-     * @param array $data
-     */
-    public function __construct(NodeFactory $nodeFactory, array $data)
+    public function render(): array
     {
-        parent::__construct($nodeFactory, $data);
+        $this->initialize();
 
-        if (isset($GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['yoast_seo'])
-            && is_array($GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['yoast_seo'])
-        ) {
-            ArrayUtility::mergeRecursiveWithOverrule(
-                $this->configuration,
-                $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['yoast_seo']
-            );
+        $resultArray = $this->initializeResultArray();
+        $resultArray['stylesheetFiles'][] = 'EXT:yoast_seo/Resources/Public/CSS/yoast.min.css';
+
+        $templateView = $this->getTemplateView();
+
+        if ($this->data['tableName'] === 'pages'
+            && !in_array((int)$this->data['databaseRow']['doktype'][0], YoastUtility::getAllowedDoktypes(), true)) {
+            $templateView->assign('wrongDoktype', true);
+            $resultArray['html'] = $templateView->render();
+            return $resultArray;
         }
 
-        $this->localeService = GeneralUtility::makeInstance(LocaleService::class, $this->configuration);
-        $this->urlService = GeneralUtility::makeInstance(UrlService::class);
+        $firstFocusKeyword = YoastUtility::getFocusKeywordOfRecord(
+            (int)$this->data['databaseRow']['uid'],
+            $this->data['tableName']
+        );
 
-        $this->templateView = GeneralUtility::makeInstance(StandaloneView::class);
-        $this->templateView->setPartialRootPaths(
-            [GeneralUtility::getFileAbsFileName('EXT:yoast_seo/Resources/Private/Partials/TCA')]
+        $snippetPreviewConfiguration = [
+            'TCA' => 1,
+            'data' => [
+                'table' => $this->data['tableName'],
+                'uid' => (int)($this->data['defaultLanguagePageRow']['uid'] ?? $this->data['databaseRow']['uid']),
+                'pid' => (int)$this->data['databaseRow']['pid'],
+                'languageId' => $this->languageId
+            ],
+            'fieldSelectors' => [
+                'title' => $this->getFieldSelector($this->titleField),
+                'pageTitle' => $this->getFieldSelector($this->pageTitleField),
+                'description' => $this->getFieldSelector($this->descriptionField),
+                'focusKeyword' => $this->getFieldSelector($this->focusKeywordField),
+                'focusKeywordSynonyms' => $this->getFieldSelector($this->focusKeywordSynonymsField),
+                'cornerstone' => $this->getFieldSelector($this->cornerstoneField),
+                'relatedKeyword' => $this->getFieldSelector($this->relatedKeyphrases, true),
+            ],
+            'relatedKeyphrases' => YoastUtility::getRelatedKeyphrases(
+                $this->data['tableName'],
+                (int)$this->data['databaseRow']['uid']
+            )
+        ];
+
+        $snippetPreviewService = GeneralUtility::makeInstance(SnippetPreviewService::class);
+        $snippetPreviewService->buildSnippetPreview(
+            $this->previewUrl,
+            $this->data['databaseRow'],
+            $snippetPreviewConfiguration
         );
-        $this->templateView->setTemplatePathAndFilename(
-            GeneralUtility::getFileAbsFileName('EXT:yoast_seo/Resources/Private/Templates/TCA/SnippetPreview.html')
-        );
+
+        $templateView->assignMultiple([
+            'previewUrl' => $this->previewUrl,
+            'previewTargetId' => $this->data['fieldName'],
+            'titleFieldSelector' => $this->getFieldSelector($this->titleField),
+            'descriptionFieldSelector' => $this->getFieldSelector($this->descriptionField),
+            'databaseRow' => $this->data['databaseRow'],
+            'focusKeyword' => $firstFocusKeyword,
+            'vanillaUid' => $this->data['vanillaUid'],
+            'tableName' => $this->data['tableName'],
+            'languageId' => $this->languageId,
+        ]);
+        $resultArray['html'] = $templateView->render();
+
+        return $resultArray;
+    }
+
+    protected function initialize(): void
+    {
+        $this->urlService = GeneralUtility::makeInstance(UrlService::class);
 
         if (array_key_exists('titleField', (array)$this->data['parameterArray']['fieldConf']['config']['settings']) &&
             $this->data['parameterArray']['fieldConf']['config']['settings']['titleField']
@@ -129,13 +106,19 @@ class SnippetPreview extends AbstractNode
             $this->titleField = $this->data['parameterArray']['fieldConf']['config']['settings']['titleField'];
         }
 
-        if (array_key_exists('pageTitleField', (array)$this->data['parameterArray']['fieldConf']['config']['settings']) &&
+        if (array_key_exists(
+                'pageTitleField',
+                (array)$this->data['parameterArray']['fieldConf']['config']['settings']
+            ) &&
             $this->data['parameterArray']['fieldConf']['config']['settings']['pageTitleField']
         ) {
             $this->pageTitleField = $this->data['parameterArray']['fieldConf']['config']['settings']['pageTitleField'];
         }
 
-        if (array_key_exists('descriptionField', (array)$this->data['parameterArray']['fieldConf']['config']['settings'])
+        if (array_key_exists(
+                'descriptionField',
+                (array)$this->data['parameterArray']['fieldConf']['config']['settings']
+            )
             && $this->data['parameterArray']['fieldConf']['config']['settings']['descriptionField']
         ) {
             $this->descriptionField =
@@ -143,8 +126,10 @@ class SnippetPreview extends AbstractNode
         }
 
         if (isset($this->data['databaseRow']['sys_language_uid'])) {
-            if (is_array($this->data['databaseRow']['sys_language_uid']) && count($this->data['databaseRow']['sys_language_uid']) > 0) {
-                $this->languageId = current($this->data['databaseRow']['sys_language_uid']);
+            if (is_array($this->data['databaseRow']['sys_language_uid']) && count(
+                    $this->data['databaseRow']['sys_language_uid']
+                ) > 0) {
+                $this->languageId = (int)current($this->data['databaseRow']['sys_language_uid']);
             } else {
                 $this->languageId = (int)$this->data['databaseRow']['sys_language_uid'];
             }
@@ -155,118 +140,19 @@ class SnippetPreview extends AbstractNode
         $this->previewUrl = $this->getPreviewUrl();
     }
 
-    public function render(): array
+    protected function getTemplateView(): StandaloneView
     {
-        $jsonConfigUtility = GeneralUtility::makeInstance(JsonConfigUtility::class);
-        $allowedDoktypes = YoastUtility::getAllowedDoktypes();
-        $resultArray = $this->initializeResultArray();
-        $resultArray['stylesheetFiles'][] = 'EXT:yoast_seo/Resources/Public/CSS/yoast.min.css';
-
-        $premiumText = $this->getPremiumText();
-
-        if ($this->data['tableName'] !== 'pages'
-            || in_array((int)$this->data['databaseRow']['doktype'][0], $allowedDoktypes)) {
-            $firstFocusKeyword = YoastUtility::getFocusKeywordOfPage(
-                (int)$this->data['databaseRow']['uid'],
-                $this->data['tableName']
-            );
-
-            $publicResourcesPath = PathUtility::getPublicPathToResources();
-            $workerUrl = $publicResourcesPath . '/JavaScript/dist/worker.js';
-
-            $config = [
-                'urls' => [
-                    'workerUrl' => $workerUrl,
-                    'previewUrl' => $this->previewUrl,
-                    'saveScores' => $this->urlService->getSaveScoresUrl(),
-                    'prominentWords' => $this->urlService->getProminentWordsUrl(),
-                ],
-                'TCA' => 1,
-                'useKeywordDistribution' => YoastUtility::isPremiumInstalled(),
-                'useRelevantWords' => YoastUtility::isPremiumInstalled(),
-                'isCornerstoneContent' => (bool)($this->data['databaseRow']['tx_yoastseo_cornerstone'] ?? false),
-                'focusKeyphrase' => [
-                    'keyword' => (string)$this->data['databaseRow']['tx_yoastseo_focuskeyword'],
-                    'synonyms' => (string)($this->data['databaseRow']['tx_yoastseo_focuskeyword_synonyms'] ?? ''),
-                ],
-                'labels' => $this->localeService->getLabels(),
-                'data' => [
-                    'table' => $this->data['tableName'],
-                    'uid' => (int)$this->data['databaseRow']['uid'],
-                    'languageId' => (int)$this->languageId
-                ],
-                'fieldSelectors' => [
-                    'title' => $this->getFieldSelector($this->titleField),
-                    'pageTitle' => $this->getFieldSelector($this->pageTitleField),
-                    'description' => $this->getFieldSelector($this->descriptionField),
-                    'focusKeyword' => $this->getFieldSelector($this->focusKeywordField),
-                    'cornerstone' => $this->getFieldSelector($this->cornerstoneField),
-                    'premiumKeyword' => YoastUtility::isPremiumInstalled() ? $this->getFieldSelector($this->relatedKeyphrases, true) : '',
-                ],
-                'translations' => $this->localeService->getTranslations(),
-                'relatedKeyphrases' => YoastUtility::isPremiumInstalled() ? YoastUtility::getRelatedKeyphrases($this->data['tableName'], (int)$this->data['databaseRow']['uid']) : []
-            ];
-
-            if (YoastUtility::isPremiumInstalled()) {
-                $config['fieldSelectors']['focusKeywordSynonyms'] = $this->getFieldSelector($this->focusKeywordSynonymsField);
-            }
-
-            $jsonConfigUtility->addConfig($config);
-
-            $pageRenderer = GeneralUtility::makeInstance(PageRenderer::class);
-
-            if (YoastUtility::inProductionMode() === true) {
-                $pageRenderer->loadRequireJsModule('TYPO3/CMS/YoastSeo/dist/plugin');
-            } else {
-                $pageRenderer->addHeaderData('<script type="text/javascript" src="https://localhost:3333/typo3conf/ext/yoast_seo/Resources/Public/JavaScript/dist/plugin.js" async></script>');
-            }
-
-            $pageRenderer->loadRequireJsModule('TYPO3/CMS/YoastSeo/yoastModal');
-
-            $this->templateView->assignMultiple([
-                'previewUrl' => $this->previewUrl,
-                'previewTargetId' => $this->data['fieldName'],
-                'titleFieldSelector' => $this->getFieldSelector($this->titleField),
-                'descriptionFieldSelector' => $this->getFieldSelector($this->descriptionField),
-                'databaseRow' => $this->data['databaseRow'],
-                'focusKeyword' => $firstFocusKeyword,
-                'vanillaUid' => $this->data['vanillaUid'],
-                'tableName' => $this->data['tableName'],
-                'languageId' => $this->languageId,
-                'previewContent' => $premiumText
-            ]);
-        } else {
-            $this->templateView->assign('wrongDoktype', true);
-        }
-        $resultArray['html'] = $this->templateView->render();
-
-        return $resultArray;
+        $templateView = GeneralUtility::makeInstance(StandaloneView::class);
+        $templateView->setPartialRootPaths(
+            [GeneralUtility::getFileAbsFileName('EXT:yoast_seo/Resources/Private/Partials/TCA')]
+        );
+        $templateView->setTemplatePathAndFilename(
+            GeneralUtility::getFileAbsFileName('EXT:yoast_seo/Resources/Private/Templates/TCA/SnippetPreview.html')
+        );
+        return $templateView;
     }
 
-    /**
-     * Get premium text
-     *
-     * @return string
-     */
-    protected function getPremiumText(): string
-    {
-        if (!YoastUtility::isPremiumInstalled()) {
-            return '
-                <div class="yoast-snippet-preview-premium">
-                    <a target="_blank" rel="noopener noreferrer" href="' . YoastUtility::getYoastLink('Go premium', 'page-properties-snippetpreview') . '">
-                        <i class="fa fa-star"></i>' . $GLOBALS['LANG']->sL('LLL:EXT:yoast_seo/Resources/Private/Language/BackendModule.xlf:goPremium') . '
-                    </a>
-                </div>';
-        }
-        return '';
-    }
-
-    /**
-     * @param string $field
-     * @param bool $id
-     * @return string
-     */
-    protected function getFieldSelector($field, $id = false): string
+    protected function getFieldSelector(string $field, bool $id = false): string
     {
         if ($id === true) {
             $element = 'data-' . $this->data['vanillaUid'] . '-' . $this->data['tableName'] . '-' . $this->data['vanillaUid'] . '-' . $field;
@@ -277,9 +163,6 @@ class SnippetPreview extends AbstractNode
         return $element;
     }
 
-    /**
-     * @return string
-     */
     protected function getPreviewUrl(): string
     {
         $currentPageId = $this->data['effectivePid'];
@@ -335,7 +218,6 @@ class SnippetPreview extends AbstractNode
         }
 
         if (!empty($previewConfiguration['useCacheHash'])) {
-            /** @var CacheHashCalculator */
             $cacheHashCalculator = GeneralUtility::makeInstance(CacheHashCalculator::class);
             $fullLinkParameters = GeneralUtility::implodeArrayForUrl(
                 '',
@@ -350,11 +232,6 @@ class SnippetPreview extends AbstractNode
         return $this->urlService->getPreviewUrl($previewPageId, $languageId, $additionalParamsForUrl);
     }
 
-    /**
-     * @param int $currentPageId
-     * @param array $previewConfiguration
-     * @return int
-     */
     protected function getPreviewPageId(int $currentPageId, array $previewConfiguration): int
     {
         // find the right preview page id
@@ -367,7 +244,7 @@ class SnippetPreview extends AbstractNode
             $currentPage = reset($rootLine);
             // Allow all doktypes below 200
             // This makes custom doktype work as well with opening a frontend page.
-            if ((int)$currentPage['doktype'] <= $this->getSpacerDoktype()) {
+            if ((int)$currentPage['doktype'] <= PageRepository::DOKTYPE_SPACER) {
                 // try the current page
                 $previewPageId = $currentPageId;
             } else {
@@ -410,24 +287,5 @@ class SnippetPreview extends AbstractNode
                 $parameters[$key] = $value;
             }
         }
-    }
-
-    /**
-     * @return int
-     */
-    protected function getSpacerDoktype(): int
-    {
-        if (class_exists(PageRepository::class)) {
-            return PageRepository::DOKTYPE_SPACER;
-        }
-        return \TYPO3\CMS\Frontend\Page\PageRepository::DOKTYPE_SPACER;
-    }
-
-    /**
-     * @return BackendUserAuthentication
-     */
-    protected function getBackendUser(): BackendUserAuthentication
-    {
-        return $GLOBALS['BE_USER'];
     }
 }
