@@ -1,11 +1,12 @@
 <?php
+
 declare(strict_types=1);
+
 namespace YoastSeoForTypo3\YoastSeo\Service;
 
 use Psr\Http\Message\UriInterface;
 use TYPO3\CMS\Backend\Routing\UriBuilder;
 use TYPO3\CMS\Backend\Utility\BackendUtility;
-use TYPO3\CMS\Core\Authentication\BackendUserAuthentication;
 use TYPO3\CMS\Core\Domain\Repository\PageRepository;
 use TYPO3\CMS\Core\Exception\SiteNotFoundException;
 use TYPO3\CMS\Core\Routing\RouteNotFoundException;
@@ -15,48 +16,27 @@ use TYPO3\CMS\Core\Site\SiteFinder;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use YoastSeoForTypo3\YoastSeo\Utility\YoastUtility;
 
-/**
- * Class UrlService
- */
 class UrlService implements SingletonInterface
 {
-    /**
-     * @var \TYPO3\CMS\Backend\Routing\UriBuilder
-     */
-    protected $uriBuilder;
+    protected UriBuilder $uriBuilder;
 
-    /**
-     * UrlService constructor.
-     */
-    public function __construct()
+    public function __construct(UriBuilder $uriBuilder)
     {
-        $this->uriBuilder = GeneralUtility::makeInstance(UriBuilder::class);
+        $this->uriBuilder = $uriBuilder;
     }
 
-    /**
-     * Get target url
-     *
-     * @param int $pageId
-     * @param int $languageId
-     * @param string $additionalGetVars
-     * @return string
-     */
     public function getPreviewUrl(
         int $pageId,
         int $languageId,
-        $additionalGetVars = ''
+        string $additionalGetVars = ''
     ): string {
         return (string)$this->uriBuilder->buildUriFromRoute('ajax_yoast_preview', [
-            'pageId' => $pageId, 'languageId' => $languageId, 'additionalGetVars' => urlencode($additionalGetVars)
+            'pageId' => $pageId,
+            'languageId' => $languageId,
+            'additionalGetVars' => urlencode($additionalGetVars)
         ]);
     }
 
-    /**
-     * @param int    $pageId
-     * @param int    $languageId
-     * @param string $additionalGetVars
-     * @return string
-     */
     public function getUriToCheck(int $pageId, int $languageId, string $additionalGetVars): string
     {
         $this->checkMountpoint($pageId, $additionalGetVars);
@@ -83,13 +63,9 @@ class UrlService implements SingletonInterface
         return '';
     }
 
-    /**
-     * @param int $pageId
-     * @param     $additionalGetVars
-     */
     public function checkMountPoint(int &$pageId, &$additionalGetVars): void
     {
-        $pageRepository = $this->getPageRepository();
+        $pageRepository = GeneralUtility::makeInstance(PageRepository::class);
         $mountPointInformation = $pageRepository->getMountPointInfo($pageId);
         if ($mountPointInformation && $mountPointInformation['overlay']) {
             // New page id
@@ -98,20 +74,11 @@ class UrlService implements SingletonInterface
         }
     }
 
-    /**
-     * @param int $pageId
-     * @return array
-     */
     public function getRootLine(int $pageId): array
     {
         return BackendUtility::BEgetRootLine($pageId);
     }
 
-    /**
-     * @param int   $pageId
-     * @param array $rootLine
-     * @return \TYPO3\CMS\Core\Site\Entity\Site|null
-     */
     public function getSite(int $pageId, array $rootLine): ?Site
     {
         $siteFinder = GeneralUtility::makeInstance(SiteFinder::class);
@@ -122,15 +89,7 @@ class UrlService implements SingletonInterface
         }
     }
 
-    /**
-     * @param \TYPO3\CMS\Core\Site\Entity\Site $site
-     * @param int                              $pageId
-     * @param int                              $languageId
-     * @param string                           $additionalGetVars
-     * @return \Psr\Http\Message\UriInterface
-     * @throws \TYPO3\CMS\Core\Routing\InvalidRouteArgumentsException
-     */
-    public function generateUri(Site $site, int $pageId, int $languageId, $additionalGetVars = ''): UriInterface
+    public function generateUri(Site $site, int $pageId, int $languageId, string $additionalGetVars = ''): UriInterface
     {
         $additionalQueryParams = [];
         $additionalGetVars = rawurldecode($additionalGetVars);
@@ -139,22 +98,6 @@ class UrlService implements SingletonInterface
         return $site->getRouter()->generateUri($pageId, $additionalQueryParams);
     }
 
-    /**
-     * @return \TYPO3\CMS\Core\Domain\Repository\PageRepository|\TYPO3\CMS\Frontend\Page\PageRepository
-     */
-    protected function getPageRepository()
-    {
-        if (class_exists(PageRepository::class)) {
-            return GeneralUtility::makeInstance(PageRepository::class);
-        }
-        return GeneralUtility::makeInstance(\TYPO3\CMS\Frontend\Page\PageRepository::class);
-    }
-
-    /**
-     * Get save scores url
-     *
-     * @return string
-     */
     public function getSaveScoresUrl(): string
     {
         try {
@@ -164,9 +107,6 @@ class UrlService implements SingletonInterface
         }
     }
 
-    /**
-     * @return string
-     */
     public function getProminentWordsUrl(): string
     {
         try {
@@ -174,13 +114,5 @@ class UrlService implements SingletonInterface
         } catch (RouteNotFoundException $e) {
             return '';
         }
-    }
-
-    /**
-     * @return \TYPO3\CMS\Core\Authentication\BackendUserAuthentication
-     */
-    protected function getBackendUser(): BackendUserAuthentication
-    {
-        return $GLOBALS['BE_USER'];
     }
 }
