@@ -1,28 +1,12 @@
 <?php
+
 declare(strict_types=1);
+
 namespace YoastSeoForTypo3\YoastSeo\DataProviders;
 
-/*
- * This file is part of the TYPO3 CMS project.
- *
- * It is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License, either version 2
- * of the License, or any later version.
- *
- * For the full copyright and license information, please read the
- * LICENSE.txt file that was distributed with this source code.
- *
- * The TYPO3 project - inspiring people to share!
- */
+use Doctrine\DBAL\Result;
 
-/**
- * Interface OverviewDataProviderInterface
- */
 interface OverviewDataProviderInterface
 {
-    /**
-     * @param bool $returnOnlyCount
-     * @return int|array
-     */
-    public function getData($returnOnlyCount = false);
+    public function getResults(array $pageIds = []): ?Result;
 }
