@@ -1,5 +1,7 @@
 <?php
+
 declare(strict_types=1);
+
 namespace YoastSeoForTypo3\YoastSeo\Form\Element;
 
 use TYPO3\CMS\Backend\Form\AbstractNode;
@@ -9,32 +11,17 @@ use TYPO3\CMS\Fluid\View\StandaloneView;
 
 class Cornerstone extends AbstractNode
 {
-    /**
-     * @var StandaloneView
-     */
-    protected $templateView;
-
-    /**
-     * @param NodeFactory $nodeFactory
-     * @param array $data
-     */
-    public function __construct(NodeFactory $nodeFactory, array $data)
-    {
-        parent::__construct($nodeFactory, $data);
-
-        $this->templateView = GeneralUtility::makeInstance(StandaloneView::class);
-        $this->templateView->setTemplatePathAndFilename(
-            GeneralUtility::getFileAbsFileName('EXT:yoast_seo/Resources/Private/Templates/TCA/Cornerstone.html')
-        );
-    }
-
     public function render(): array
     {
         $resultArray = $this->initializeResultArray();
 
-        $this->templateView->assign('data', $this->data);
+        $templateView = GeneralUtility::makeInstance(StandaloneView::class);
+        $templateView->setTemplatePathAndFilename(
+            GeneralUtility::getFileAbsFileName('EXT:yoast_seo/Resources/Private/Templates/TCA/Cornerstone.html')
+        );
+        $templateView->assign('data', $this->data);
 
-        $resultArray['html'] = $this->templateView->render();
+        $resultArray['html'] = $templateView->render();
 
         return $resultArray;
     }
