@@ -1,27 +1,31 @@
 <?php
+
 declare(strict_types=1);
+
 namespace YoastSeoForTypo3\YoastSeo\Form\Element;
 
 use TYPO3\CMS\Backend\Form\AbstractNode;
-use TYPO3\CMS\Backend\Form\NodeFactory;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Fluid\View\StandaloneView;
 
 class Insights extends AbstractNode
 {
-    /**
-     * @param NodeFactory $nodeFactory
-     * @param array $data
-     */
-    public function __construct(NodeFactory $nodeFactory, array $data)
-    {
-        parent::__construct($nodeFactory, $data);
-    }
-
     public function render(): array
     {
         $resultArray = $this->initializeResultArray();
 
-        $resultArray['html'] = '<div><a href="#" data-yoast-modal-type="insights" class="btn btn-default">Get insights</a></div>';
+        $templateView = GeneralUtility::makeInstance(StandaloneView::class);
+        $templateView->setPartialRootPaths(
+            [GeneralUtility::getFileAbsFileName('EXT:yoast_seo/Resources/Private/Partials/TCA')]
+        );
+        $templateView->setTemplatePathAndFilename(
+            GeneralUtility::getFileAbsFileName(
+                'EXT:yoast_seo/Resources/Private/Templates/TCA/Insights.html'
+            )
+        );
 
+        $templateView->assign('data', $this->data);
+        $resultArray['html'] = $templateView->render();
         return $resultArray;
     }
 }
