@@ -1,5 +1,7 @@
 <?php
+
 declare(strict_types=1);
+
 namespace YoastSeoForTypo3\YoastSeo\Middleware;
 
 use Psr\Http\Message\ResponseInterface;
@@ -11,18 +13,8 @@ use TYPO3\CMS\Core\Context\VisibilityAspect;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use YoastSeoForTypo3\YoastSeo\Utility\YoastRequestHash;
 
-/**
- * Class PageRequestMiddleware
- */
 class PageRequestMiddleware implements MiddlewareInterface
 {
-    /**
-     * Process page request
-     *
-     * @param \Psr\Http\Message\ServerRequestInterface $request
-     * @param \Psr\Http\Server\RequestHandlerInterface $handler
-     * @return \Psr\Http\Message\ResponseInterface
-     */
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
         if (YoastRequestHash::isValid($request->getServerParams())) {
