@@ -1,12 +1,10 @@
 <?php
+
 declare(strict_types=1);
 
 namespace YoastSeoForTypo3\YoastSeo\StructuredData;
 
 interface StructuredDataProviderInterface
 {
-    /**
-     * @return array
-     */
     public function getData(): array;
 }
