@@ -5,39 +5,81 @@ declare(strict_types=1);
 namespace YoastSeoForTypo3\YoastSeo\Utility;
 
 use TYPO3\CMS\Core\Authentication\BackendUserAuthentication;
-use TYPO3\CMS\Core\Database\QueryGenerator;
+use TYPO3\CMS\Core\Database\ConnectionPool;
+use TYPO3\CMS\Core\Database\Query\QueryHelper;
+use TYPO3\CMS\Core\Database\Query\Restriction\DeletedRestriction;
 use TYPO3\CMS\Core\Type\Bitmask\Permission;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 class PageAccessUtility
 {
-    protected static $cache = [];
+    protected static array $cache = [];
 
-    public static function getPageIds(): ?array
+    public static function getPageIds(int $pid): array
     {
-        $backendUser = self::getBackendUser();
-        if ($backendUser->isAdmin()) {
-            return null;
-        }
-
         if (self::$cache !== []) {
             return self::$cache;
         }
 
-        $webMounts = $backendUser->returnWebmounts();
+        $backendUser = self::getBackendUser();
         $perms_clause = $backendUser->getPagePermsClause(Permission::PAGE_SHOW);
 
-        $pageIds = '';
-        foreach ($webMounts as $mount) {
-            $pageIds .= (!empty($pageIds) ? ',' : '') . self::getTreeList((int)$mount, $perms_clause);
-        }
-
-        return self::$cache = GeneralUtility::intExplode(',', $pageIds);
+        return self::$cache = GeneralUtility::intExplode(
+            ',',
+            self::getTreeList($pid, 999, 0, $perms_clause)
+        );
     }
 
-    protected static function getTreeList(int $pid, string $perms_clause): string
+    /**
+     * Copied from EXT:core, removed in v12
+     *
+     * @param $id
+     * @param $depth
+     * @param $begin
+     * @param $permClause
+     * @return string
+     */
+    protected static function getTreeList(int $id, int $depth, int $begin = 0, string $permClause = ''): string
     {
-        return (string)GeneralUtility::makeInstance(QueryGenerator::class)->getTreeList($pid, 999, 0, $perms_clause);
+        $depth = (int)$depth;
+        $begin = (int)$begin;
+        $id = (int)$id;
+        if ($id < 0) {
+            $id = abs($id);
+        }
+        if ($begin === 0) {
+            $theList = $id;
+        } else {
+            $theList = '';
+        }
+        if ($id && $depth > 0) {
+            $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable('pages');
+            $queryBuilder->getRestrictions()->removeAll()->add(GeneralUtility::makeInstance(DeletedRestriction::class));
+            $queryBuilder->select('uid')
+                ->from('pages')
+                ->where(
+                    $queryBuilder->expr()->eq('pid', $queryBuilder->createNamedParameter($id, \PDO::PARAM_INT)),
+                    $queryBuilder->expr()->eq('sys_language_uid', 0)
+                )
+                ->orderBy('uid');
+            if ($permClause !== '') {
+                $queryBuilder->andWhere(QueryHelper::stripLogicalOperatorPrefix($permClause));
+            }
+            $statement = $queryBuilder->execute();
+            while ($row = $statement->fetchAssociative()) {
+                if ($begin <= 0) {
+                    $theList .= ',' . $row['uid'];
+                }
+                if ($depth > 1) {
+                    $theSubList = self::getTreeList($row['uid'], $depth - 1, $begin - 1, $permClause);
+                    if (!empty($theList) && !empty($theSubList) && ($theSubList[0] !== ',')) {
+                        $theList .= ',';
+                    }
+                    $theList .= $theSubList;
+                }
+            }
+        }
+        return (string)$theList;
     }
 
     protected static function getBackendUser(): BackendUserAuthentication
