@@ -1,5 +1,7 @@
 <?php
+
 declare(strict_types=1);
+
 namespace YoastSeoForTypo3\YoastSeo\Utility;
 
 use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
@@ -9,6 +11,8 @@ class PathUtility
 {
     public static function getPublicPathToResources(): string
     {
-        return TYPO3PathUtility::getAbsoluteWebPath(ExtensionManagementUtility::extPath('yoast_seo') . 'Resources/Public');
+        return TYPO3PathUtility::getAbsoluteWebPath(
+            ExtensionManagementUtility::extPath('yoast_seo') . 'Resources/Public'
+        );
     }
 }
