@@ -2,6 +2,7 @@ import {Paper} from 'yoastseo';
 import measureTextWidth from '../helpers/measureTextWidth';
 import {analyzeData} from '../redux/actions/analysis';
 import {getRelevantWords} from '../redux/actions/relevantWords';
+import {getInsights} from "../redux/actions/insights";
 
 export default function refreshAnalysis(worker, store) {
     const state = store.getState();
@@ -16,18 +17,9 @@ export default function refreshAnalysis(worker, store) {
         titleWidth: measureTextWidth(data.title)
     });
 
-    let promises = [];
-    if (typeof YoastConfig.useRelevantWords !== 'undefined' &&
-        YoastConfig.useRelevantWords === true) {
-        promises = [
-            store.dispatch(analyzeData(worker, paper, YoastConfig.relatedKeyphrases)),
-            store.dispatch(getRelevantWords(worker, paper)),
-        ];
-    } else {
-        promises = [
-            store.dispatch(analyzeData(worker, paper, YoastConfig.relatedKeyphrases)),
-        ];
-    }
-
-    return Promise.all(promises);
+    return Promise.all([
+        store.dispatch(analyzeData(worker, paper, YoastConfig.relatedKeyphrases)),
+        store.dispatch(getRelevantWords(worker, paper)),
+        store.dispatch(getInsights(worker, paper))
+    ]);
 }
