@@ -3,11 +3,15 @@ import {setFocusKeyword} from './focusKeyword';
 import {setFocusKeywordSynonyms} from './focusKeywordSynonyms';
 import {analyzeData} from "./analysis";
 import {getRelevantWords} from "./relevantWords";
+import {getInsights} from "./insights";
+import {getLinkingSuggestions} from "./linkingSuggestions";
 
 export default {
     getContent,
     setFocusKeyword,
     setFocusKeywordSynonyms,
     analyzeData,
-    getRelevantWords
+    getRelevantWords,
+    getInsights,
+    getLinkingSuggestions
 };
