@@ -3,11 +3,14 @@ import focusKeyword from './focusKeyword';
 import focusKeywordSynonyms from './focusKeywordSynonyms';
 import analysis from './analysis';
 import relevantWords from './relevantWords';
-
+import insights from "./insights";
+import linkingSuggestions from "./linkingSuggestions";
 export default {
     content,
     focusKeyword,
     focusKeywordSynonyms,
     analysis,
-    relevantWords
+    relevantWords,
+    insights,
+    linkingSuggestions
 }
