@@ -1,11 +1,8 @@
 module.exports = {
     build: {
         files: {
-            'Resources/Public/CSS/yoast-seo.min.css': 'Resources/Private/SCSS/YoastTYPO3.scss',
-            'Resources/Public/CSS/yoast-seo-tca.min.css': 'Resources/Private/SCSS/YoastTCA.scss',
-            'Resources/Public/CSS/yoast-seo-snippet-preview.min.css': 'Resources/Private/SCSS/YoastSnippetPreview.scss',
-            'Resources/Public/CSS/yoast-seo-backend.min.css': 'Resources/Private/SCSS/BackendModule.scss',
-            'Resources/Public/CSS/yoast.min.css': 'Resources/Private/SCSS/Yoast.scss',
+            './../Resources/Public/CSS/yoast-seo-backend.min.css': './../Resources/Private/SCSS/BackendModule.scss',
+            './../Resources/Public/CSS/yoast.min.css': './../Resources/Private/SCSS/Yoast.scss',
         }
     }
 };
